bgituQuizzAdmin/requirements.txt
Django>=4.2
djangorestframework
python-dotenv
psycopg2-binary
requests
unfold 